cargo: support `serde_yaml` 0.9
The 0.9 release adds support for tagged values. Since the core
`yaml-rust` crate doesn't support them, they are instead smuggled
through using a hash with two keys to carry the tag and value
separately. This does mean that there could be values which turn into
tagged values when round-tripping, but the hashes must meet the exact
requirements which is considered to be suitable.

Also updates the minimum Rust to 1.56 to support the 2021 edition used
in the now-minimum `lazy_static`.

Fixes: #6